Creditcard Validation
This Java program defines a Creditcard class that validates credit card numbers. It checks if the credit card number is between 8 and 9 digits long and removes the last digit for validation. The program then reverses the remaining digits and doubles the values at odd positions, summing them to check against the last digit. The card is valid if the calculated expected digit matches the extracted last digit , otherwise we say card is invalid.