AlphabetWarGame
This Java program defines an AlphabetWarGame class that sees a war between two sides based on the letters given in a word. The program assigns weightage to specific letters for left and right side. Using this weightage the program determines with the user input given , whether there is a winner (left or right side) or should they start again.